Abstracting socket handler out from the main app
